fixed layout of the account creation frame
