+++ src/UI/Profiles.form
@@ -33,13 +33,20 @@
         <border type="none"/>
         <children/>
       </grid>
-      <grid id="c44ca" binding="test3" layout-manager="FlowLayout" hgap="5" vgap="5" flow-align="1">
+      <grid id="c44ca" binding="test3" layout-manager="BorderLayout" hgap="0" vgap="0">
         <constraints border-constraint="South"/>
         <properties>
           <opaque value="false"/>
         </properties>
         <border type="none"/>
-        <children/>
+        <children>
+          <component id="da350" class="javax.swing.JButton" binding="backButton">
+            <constraints border-constraint="West"/>
+            <properties>
+              <text value="&lt;-- Back"/>
+            </properties>
+          </component>
+        </children>
       </grid>
       <grid id="35afd" binding="middle" layout-manager="GridBagLayout">
         <constraints border-constraint="Center"/>
